rip out old soulbind logic, replace with new
How Soulbind Charts worked so far:
- Simulate a profile without any soulbind effects (the covenant is still active). This is the baseline dps of a Covenant.
- Simulate all Soulbind effects separately. Each simulated profile is compared to the appropriate Covenant baseline dps.
- Simulate all single/double Conduits & Ranks. Each simulated profile is compared to the appropriate Covenant baseline dps.
- Figure out all potential paths through a Soulbind tree. Sum up the previously generated dps gains for all paths. The winner represents the highest gain for each Soulbind.

The issue:
This way synergies between Soulbind effects & Soulbind effects, Soulbind effects & Conduits, and Conduits & Conduits are not part of the results. An additional issue can arise if a spec has some sort of very specific goal, like Elemental Shamans reaching 100% uptime on their Storm Elemental. Checking only single effects might result in heavy under- or even overvaluing of effects. With no direct way of knowing.

Going forward:
Soulbind Chart will work differently in the future.
- For each Soulbind all potential unique dps paths are simulated. Including Conduits at max rank.
- The chart informs about the winning path of each Soulbind.
- This means you can see how different Soulbinds perform and which effects in combination yielded the highest dps.

Drawback:
The value of a single effect is no longer visible. There may or may not come a chart for that. But because single effects worth can be highly dependant on the presence of other effects, this is not set in stone for now.